% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_topics.R
\name{get_category_topics}
\alias{get_category_topics}
\title{Returns the topics for a given category}
\usage{
get_category_topics(category)
}
\arguments{
\item{category}{Valid discourse category name}
}
\description{
Returns the topics for a given category
}
