% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{get_group_members}
\alias{get_group_members}
\title{Returns the member usernames from a given group}
\usage{
get_group_members(groupname)
}
\arguments{
\item{groupname}{Valid discourse group name}
}
\description{
Returns the member usernames from a given group
}
