% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user_info}
\alias{get_user_info}
\title{Gets useful user information}
\usage{
get_user_info(username)
}
\arguments{
\item{username}{A valid discourse username}
}
\description{
Gets useful user information
}
